Feed Fish, folder restructure
Implemented the functionality for feeding a fish including instantiating flakes, having the fish check for flakes if they're hungry, move towards the flake to eat it, start an eating rampage and only send a request to the API at the end of the rampage with the increase in hunger points. Also restructured the folders greatly because stuff was getting messy with more files... Also implemented a FishFlip script for flipping its sprite left or right since that was an issue. Cheerzzz!!!!