providing dead outlines in fishController and also cursor resize
included changes to how the cursor works - before it worked with a series of Textures (separate images) but i needed to change the SIZE of the texture and this wasn't working well programatically, trust me i tried... so i made it so that now it takes Sprites (which are far easier to generate from a sprite sheet, i wasn't about to manually make 5 images every time i wanna change an animation) and it converts them to Textures inside the code - everything else remained the same, so i basically just upscaled the image and reintroduced it in the project files (it's 2x as big as before and it looks great) i also added a sprite sheet that's 4x as big as before just in case it's useful sometime in the future. also sped up the petting animation a bit and it's more satisfying now. :D