knocking functionalityyy
in summary what happens is:

a new GameObject in the scene called Knocker, and a new Component in the Fish prefab called FishKnockReaction - all of the FishKnockReaction have a pointer to Knocker because they become its listeners! so actually when you double tap, the Knocker sends out an event with the position of the tap :). and the Fish listen to that event and check if it was close enough for them to react. If it's close enough, they swim away from the knock (so if you do it in the middle of a bunch of fish, they swim away radially it's awesome). The code about "swimming away" was added in FishMovement which was renamed from FishIdleSwim. The Knocker also has a reference to the PettingManager (cursor manager) to change the cursor to the knocking one. So of course also added a new Cursor type :) but yeah that's pretty much it